Undo p2p changes in this PR
